% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.nl_cache_key}
\alias{.nl_cache_key}
\title{Generate cache key from inputs}
\usage{
.nl_cache_key(...)
}
\arguments{
\item{...}{Named arguments to hash}
}
\value{
SHA-256 cache key
}
\description{
Generate cache key from inputs
}
\keyword{internal}
