% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.nl_cache_set}
\alias{.nl_cache_set}
\title{Set cached result}
\usage{
.nl_cache_set(key, value)
}
\arguments{
\item{key}{Cache key from .nl_cache_key()}

\item{value}{Value to cache}
}
\description{
Set cached result
}
\keyword{internal}
