% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{.nl_get_engine}
\alias{.nl_get_engine}
\title{Get a custom engine}
\usage{
.nl_get_engine(name)
}
\arguments{
\item{name}{Engine name}
}
\value{
List with fit, predict, validate functions, or NULL
}
\description{
Get a custom engine
}
\keyword{internal}
