% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{.nl_has_engine}
\alias{.nl_has_engine}
\title{Check if a custom engine is registered}
\usage{
.nl_has_engine(name)
}
\arguments{
\item{name}{Engine name}
}
\value{
Logical
}
\description{
Check if a custom engine is registered
}
\keyword{internal}
