% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{.nl_safe_vars}
\alias{.nl_safe_vars}
\title{Get safe variable names that exist in data}
\usage{
.nl_safe_vars(object, type = c("predictor", "response"))
}
\arguments{
\item{object}{An nlrfit or data frame}

\item{type}{"predictor" or "response"}
}
\value{
Character vector of variable names present in data
}
\description{
Get safe variable names that exist in data
}
\keyword{internal}
