% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{.nl_validate_group}
\alias{.nl_validate_group}
\title{Validate group column}
\usage{
.nl_validate_group(object, group = NULL)
}
\arguments{
\item{object}{An nlrfit}

\item{group}{Group name or NULL}
}
\value{
Validated group name or NULL
}
\description{
Validate group column
}
\keyword{internal}
