% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{.nl_validate_nlrfit}
\alias{.nl_validate_nlrfit}
\title{Validate nlrfit object}
\usage{
.nl_validate_nlrfit(object)
}
\arguments{
\item{object}{Object to validate}
}
\value{
TRUE invisibly
}
\description{
Validate nlrfit object
}
\keyword{internal}
