% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{nl_list_engines}
\alias{nl_list_engines}
\title{List registered engines}
\usage{
nl_list_engines()
}
\value{
Character vector of engine names
}
\description{
Returns the names of all registered custom engines.
}
