% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{nlrFlow-package}
\alias{nlrFlow-package}
\alias{nlrFlow}
\title{nlrFlow: integrated scientific workflows for nonlinear regression}
\description{
The generics that nlrFlow registers S3 methods for live in \pkg{stats}, so
they are imported explicitly; otherwise \code{loadNamespace()} would fail while
registering, for example, \code{S3method(BIC, nlrfit)} when only the base
namespace is loaded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wep69/nlrFlow}
  \item Report bugs at \url{https://github.com/wep69/nlrFlow/issues}
}

}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
